Revert "Fix version compare between prerelease and stable"
It's not the right way because it's not how wpkg works internally.
Now there is a fix with wpkg and the comparer in case of versions
like:

1.48~rc1
1.48

where (with tilde), 1.48 is greater than 1.48~rc1

It means that you must use tidle [~] instead of minus [-] with the
prelreleases.

See https://windowspackager.org/documentation/implementation-details/debian-version

This reverts commit 9abea83fdf04f2ade3ecb09b613750a8c2efa6d8.